Fix #36: incorrect parsing of large unsigned decimal literals
All text formats -- Piq, Json and XML -- had this problem as they all relied on
the same buggy primitive for parsing unsigned decimals > 2^63 - 1

+++ piqilib/piqi_c_impl.c
@@ -49,7 +49,7 @@ value camlidl_piqi_c_piqi_strtoull(
 	value _v_str)
 {
   char *str; /*in*/
-  int _res;
+  int64 _res;
   value _vres;
 
   str = String_val(_v_str);
